Refactor Apify webhook handler to support multiple actor types
